split parts of the scatter index:
- colorbar and handleXYDefaults get their own files.
